requirements.txt
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
pandas>=2.2
apscheduler>=3.10
streamlit>=1.32
pyyaml>=6.0
python-dotenv>=1.0
truststore>=0.10
pytest>=8.0
